fast-boot: Fixing shuffle code and testing documentation
Impact: Silence the output

Fix some bugs in the shuffling code.

Silence the output. DEBUG is used for critical information to check the
parser does its job. METADATA_DEBUG is really, really verbose.

Add documentation to test the SYSLINUX code using a Virtual Machine and
VMware Workstation.

Testing Done: used resume.c32 in a VM to jump into protected mode. I
needed to increase the RAM of the VM (the default 256MB is way to
small).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>