fast-boot: Initial implementation of the trampoline
Impact: extract inline asm to real assembly
	break temporary the parser1 due to a bug in core

This implements the main assembly code for restoring the saved state.
Although it successfully jumps into PM, it triple faults soon after.
This commits is simply to introduce the final trampoline infrastructure.

Due to a potential bug in SYSLINUX with 681 segments to shuffle, let's
stop the parser after 680 segments. This needs to be fixed though.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>