fast-boot: Relocate trampoline at __nosave_begin
Impact: Change relocation scheme.

Fix relocation in the trampoline. We actually need to relocate ourselves
in two different places.

The trampoline needs to be in the kernel address space (so that the
code is mapped via swapper_pg_dir). But the area
	__nosave_begin...__nosave_end
is too small for code + data (our boot page tables are huge). So we
relocate the data in low memory (0x8000).

Testing Done: compared values of saved_context_* restored to the same
ones in the kernel (via printks). They all match. Checked cr3 is loaded
with the physical value present in the System.map.

There is one outstanding bug: the stack is not properly mapped by
swapper_pg_dir (pop/push generate a page fault).

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>