fast-boot: Restore extra symbols needed by TuxOnIce or swsusp
Impact: new symbols restored

TuxOnIce and swsusp mark some symbols as nosave (which are stored
between __nosave_begin and __nosave_end). They need to be restored as
well.
There is still one (restore_highmem_pblist) that we don't restore. It is
used by TuxOnIce when aborting a cycle. Unfortunately we can't support
aborting (a full reboot is necessary) - so we can skip it.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>