hdt: Add LZO support, drop LZF
Impact: resume.c32 is only compatible with TuxOnIce 3.0+

Upstream has switched to LZO which is in linus tree (this
was not the case with LZF).

We use the miniLZO library:

  Author  : Markus Franz Xaver Johannes Oberhumer
            <markus@oberhumer.com>
            http://www.oberhumer.com/opensource/lzo/
  Version : 2.03
  Date    : 30 Apr 2008

Let's put the library in com32/resume/lzo for now to limit the impact of
the resume.c32 module. If it is upstreamed, this will need more
thinking.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>