fast-boot: Smart image parsing
Impact: major code refactoring

Implement "smart" image parsing to autodetect the method to use
to resume (swsusp or TuxOnIce). This is done via a new structure
resume_info that tracks global parameters.

resume.c32 is now capable of detecting if a swsusp hibernate
image is present on a swap partition.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>